LayoutTests/fast/js/Promise-already-fulfilled.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;

new Promise(function(resolver) {
  resolver.fulfill('foo');
  resolver.fulfill('fulfill');
  resolver.resolve('resolve');
  resolver.reject('reject');
}).then(function(result) {
  window.result = result;
  shouldBeEqualToString('result', 'foo');
  finishJSTest();
}, function() {
  testFailed('rejected');
  finishJSTest();
});

</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
